Train each rating dimension and reconstruct full featurized dataframe
This commit encapsulates the full featurization of the dataset based on the 5 rating dimensions learned on IMDB's comments. The target has been pushed to S3 after generation in doc2vec_code/. The process is as follow, for each dimension in ['VIOLENCE', 'DRUGS', 'NUDITY', 'PROFANITY', 'SCARY'].
- Split bag of words between labeled and unlabeled dataset
- Train (n-1) folds on labeled dataset
- Predict probability on n-th fold
- Predict probability for all unlabeled observations
- Repeat n times
Then:
- Concatenate all n-th folds to reconstruct feature on the whole labeled dataset
- Average probabilities on unlabeled observations
- Concatenate labeled and unlabeled prediction
Finally, repeat for each category, and join each newly engineered feature into final dataframe to be saved to disk.